# internal/contract/testdata/openapi.baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Regnovum API",
    "version": "0.1.0",
    "description": "Public HTTP contract of Regnovum. Versioned per the master plan: routes registered by the binary must match this document; the contract test fails on drift.",
    "summary": "Arena API contract"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API surface served by the arena binary."
    }
  ],
  "paths": {
...
    "/d/{slug}": {
      "get": {
        "operationId": "getArenaDocument",
        "summary": "Public Arena document (HTML)",
        "description": "Renders the cacheable public HTML document of an Arena addressed by its slug: title, description, canonical link, OpenGraph and JSON-LD built from the Arena's own content and content_language. Metadata never includes aggregated results (none exist in the MVP). Removed Arenas answer 410 Gone and every other unresolvable value answers 404; error documents are never cached and never echo the requested address.",
        "tags": ["arenas"],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Stable public Arena address.",
            "schema": { "type": "string", "minLength": 3, "maxLength": 80 }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Validator returned by a previous read; a match answers 304 without a body.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Public Arena document.",
            "headers": {
              "ETag": { "description": "Strong validator of the exact response document; revalidate with If-None-Match.", "schema": { "type": "string" } },
              "Cache-Control": { "description": "Public cache policy (public, max-age=60).", "schema": { "type": "string" } }
            },
            "content": {
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          },
          "304": { "description": "Not modified: the If-None-Match validator matches the current document." },
          "404": {
            "description": "Arena not found. The error document is not cacheable.",
            "content": {
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          },
          "410": {
            "description": "Arena removed by moderation. The error document is not cacheable.",
            "content": {
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arena_session",
        "description": "Opaque session token cookie (Secure, HttpOnly, SameSite). Sessions arrive with the identity phase; the scheme is declared now to anchor the contract."
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Per-request CSRF token required by state-changing requests once sessions exist."
      }
    },
...
  },
...
  "security": [],
...
}